# docs/swagger.yaml
definitions:
  Global-Response:
    properties:
      message: {}
      result: {}
      status:
        type: boolean
    type: object
  moviecontroller.StoreJson:
    properties:
      artists:
        type: string
      description:
        type: string
      duration:
        type: integer
      genres:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  movievotecontroller.VoteRequest:
    properties:
      id_movie:
        type: integer
    type: object
  usercontroller.LoginRequestJson:
    properties:
      password:
        example: edp
        type: string
      username:
        example: edp
        type: string
    required:
    - password
    - username
    type: object
  usercontroller.RegisterRequestJson:
    properties:
      password:
        example: edp
        type: string
      username:
        example: edp
        type: string
    required:
    - password
    - username
    type: object
info:
  contact: {}
  title: Service message docs
  version: 1.0.0
paths:
  /login:
    post:
      consumes:
      - application/json
      description: login service account
      parameters:
      - description: body
        in: body
        name: raw
        required: true
        schema:
          $ref: '#/definitions/usercontroller.LoginRequestJson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Global-Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Global-Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Global-Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Global-Response'
      summary: login
      tags:
      - Account
  /movies:
    get:
      consumes:
      - application/json
      description: Movie service
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: title
        in: query
        name: title
        type: string
      - description: description
        in: query
        name: description
        type: string
      - description: artists
        in: query
        name: artists
        type: string
      - description: genres
        in: query
        name: genres
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Global-Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Global-Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Global-Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Global-Response'
      summary: Movie
      tags:
      - Movies
    post:
      consumes:
      - application/json
      description: Movie service
      parameters:
      - description: body
        in: body
        name: raw
        required: true
        schema:
          $ref: '#/definitions/moviecontroller.StoreJson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Global-Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Global-Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Global-Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Global-Response'
      security:
      - Bearer: []
      summary: Movie
      tags:
      - Movies
  /movies/{id}:
    put:
      consumes:
      - application/json
      description: Movie service
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: raw
        required: true
        schema:
          $ref: '#/definitions/moviecontroller.StoreJson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Global-Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Global-Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Global-Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Global-Response'
      security:
      - Bearer: []
      summary: Movie
      tags:
      - Movies
  /movies/unvote:
    post:
      consumes:
      - application/json
      description: Unvote Movie service
      parameters:
      - description: body
        in: body
        name: raw
        required: true
        schema:
          $ref: '#/definitions/movievotecontroller.VoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Global-Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Global-Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Global-Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Global-Response'
      security:
      - Bearer: []
      summary: Unvote Movie
      tags:
      - Movies-Vote
  /movies/vote:
    post:
      consumes:
      - application/json
      description: Vote Movie service
      parameters:
      - description: body
        in: body
        name: raw
        required: true
        schema:
          $ref: '#/definitions/movievotecontroller.VoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Global-Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Global-Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Global-Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Global-Response'
      security:
      - Bearer: []
      summary: Vote movie
      tags:
      - Movies-Vote
  /register:
    post:
      consumes:
      - application/json
      description: register
      parameters:
      - description: body
        in: body
        name: raw
        required: true
        schema:
          $ref: '#/definitions/usercontroller.RegisterRequestJson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Global-Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Global-Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Global-Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Global-Response'
      summary: register
      tags:
      - Account
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"